+++ src/core/provider/index.ts
@@ -1,0 +1,1 @@
+export { RainbowProvider } from './RainbowProvider';
