+++ src/entries/background/index.ts
