+++ src/entries/popup/index.ts
@@ -1,7 +1,8 @@
 import { createElement } from 'react';
 import ReactDOM from 'react-dom';
-import { Popup } from '~/pages/popup';
-import '~/styles/global.css';
+
+import { Index } from './pages';
+import './global.css';
 
 const domContainer = document.querySelector('#app');
-ReactDOM.render(createElement(Popup), domContainer);
+ReactDOM.render(createElement(Index), domContainer);
