+++ src/entries/popup/pages/index.css.ts
